Chat fixed

+++ app/controllers/chat_messages_controller.rb
@@ -3,12 +3,17 @@ class ChatMessagesController < InheritedResources::Base
   respond_to :json
 
   def create
-    @chat_message = ChatMessage.new :message => params[:message],
-                                    :user => current_user
-    @chat_message.save
-    if ChatMessage.count > 50
-      ChatMessage.first.destroy
+    last = ChatMessage.last
+
+    unless last.message == params[:message] and last.user == current_user
+      @chat_message = ChatMessage.new :message => params[:message],
+                                      :user => current_user
+      @chat_message.save
+      if ChatMessage.count > 50
+        ChatMessage.first.destroy
+      end
     end
+
     render 'show'
   end
 
@@ -17,6 +22,10 @@ class ChatMessagesController < InheritedResources::Base
       @chat_messages = ChatMessage.where("id > ?", params[:id])
     end
     render 'index'
+  end
+
+  def index
+    @chat_messages = ChatMessage.all.reorder(:id)
   end
 
   private
